requirements.txt
Django==6.0.3
djangorestframework
cryptography
webauthn
